Replace int/uint by isze/usize

+++ src/allocator.rs
@@ -12,27 +12,27 @@ pub trait Allocator {
     // FIXME: (#16293) waiting UFCS to remove this unused first argument.
     /// Allocate `size` bytes of memory whose base address is a multiple
     /// of `align`.
-    unsafe fn allocate(_: Option<Self>, size: uint, align: uint) -> *mut u8;
+    unsafe fn allocate(_: Option<Self>, size: usize, align: usize) -> *mut u8;
 
     /// Deallocate `size` bytes memory at `ptr`. `size` and `align` must
     /// be the same values used when `allocate` was called.
-    unsafe fn deallocate(_: Option<Self>, ptr: *mut u8, size: uint,
-                         align: uint);
+    unsafe fn deallocate(_: Option<Self>, ptr: *mut u8, size: usize,
+                         align: usize);
 }
 
 /// Trait for keys allocators
 pub trait KeyAllocator : Allocator {
     /// Set memory protection of pages allocated at `ptr` to read-only.
-    unsafe fn protect_read(_: Option<Self>, ptr: *mut u8, size: uint);
+    unsafe fn protect_read(_: Option<Self>, ptr: *mut u8, size: usize);
 
     /// Set memory protection of pages allocated at `ptr` to write-only.
     /// Note that on most architectures write-only cannot be enforced
     /// and also allows read access.
-    unsafe fn protect_write(_: Option<Self>, ptr: *mut u8, size: uint);
+    unsafe fn protect_write(_: Option<Self>, ptr: *mut u8, size: usize);
 
     /// Set memory protection of pages allocated at `ptr` to prevent
     /// any access.
-    unsafe fn protect_none(_: Option<Self>, ptr: *mut u8, size: uint);
+    unsafe fn protect_none(_: Option<Self>, ptr: *mut u8, size: usize);
 }
 
 
@@ -59,28 +59,28 @@ pub type DefaultKeyAllocator = ProtectedKeyAllocator;
 pub struct NullHeapAllocator;
 
 impl Allocator for NullHeapAllocator {
-    unsafe fn allocate(_: Option<NullHeapAllocator>, size: uint,
-                       align: uint) -> *mut u8 {
+    unsafe fn allocate(_: Option<NullHeapAllocator>, size: usize,
+                       align: usize) -> *mut u8 {
         heap::allocate(size, align)
     }
 
     unsafe fn deallocate(_: Option<NullHeapAllocator>, ptr: *mut u8,
-                         size: uint, align: uint) {
+                         size: usize, align: usize) {
         heap::deallocate(ptr, size, align);
     }
 }
 
 impl KeyAllocator for NullHeapAllocator {
     unsafe fn protect_read(_: Option<NullHeapAllocator>, _ptr: *mut u8,
-                           _size: uint) {
+                           _size: usize) {
     }
 
     unsafe fn protect_write(_: Option<NullHeapAllocator>, _ptr: *mut u8,
-                            _size: uint) {
+                            _size: usize) {
     }
 
     unsafe fn protect_none(_: Option<NullHeapAllocator>, _ptr: *mut u8,
-                           _size: uint) {
+                           _size: usize) {
     }
 }
 
@@ -94,13 +94,13 @@ impl KeyAllocator for NullHeapAllocator {
 pub struct ProtectedBufferAllocator;
 
 impl Allocator for ProtectedBufferAllocator {
-    unsafe fn allocate(_: Option<ProtectedBufferAllocator>, size: uint,
-                       align: uint) -> *mut u8 {
+    unsafe fn allocate(_: Option<ProtectedBufferAllocator>, size: usize,
+                       align: usize) -> *mut u8 {
         malloc::malloc(size, align)
     }
 
     unsafe fn deallocate(_: Option<ProtectedBufferAllocator>, ptr: *mut u8,
-                         _size: uint, _align: uint) {
+                         _size: usize, _align: usize) {
         malloc::free(ptr);
     }
 }
@@ -115,30 +115,30 @@ impl Allocator for ProtectedBufferAllocator {
 pub struct ProtectedKeyAllocator;
 
 impl Allocator for ProtectedKeyAllocator {
-    unsafe fn allocate(_: Option<ProtectedKeyAllocator>, size: uint,
-                       align: uint) -> *mut u8 {
+    unsafe fn allocate(_: Option<ProtectedKeyAllocator>, size: usize,
+                       align: usize) -> *mut u8 {
         malloc::malloc_key(size, align)
     }
 
     unsafe fn deallocate(_: Option<ProtectedKeyAllocator>, ptr: *mut u8,
-                         _size: uint, _align: uint) {
+                         _size: usize, _align: usize) {
         malloc::free(ptr);
     }
 }
 
 impl KeyAllocator for ProtectedKeyAllocator {
     unsafe fn protect_read(_: Option<ProtectedKeyAllocator>, ptr: *mut u8,
-                           _size: uint) {
+                           _size: usize) {
         malloc::protect_read(ptr);
     }
 
     unsafe fn protect_write(_: Option<ProtectedKeyAllocator>,
-                            ptr: *mut u8, _size: uint) {
+                            ptr: *mut u8, _size: usize) {
         malloc::protect_write(ptr);
     }
 
     unsafe fn protect_none(_: Option<ProtectedKeyAllocator>, ptr: *mut u8,
-                           _size: uint) {
+                           _size: usize) {
         malloc::protect_none(ptr);
     }
 }
